CMake init
