+++ src/jacobi.cpp
@@ -3,7 +3,7 @@
 
 // если это делать по нормальному, то у меня компилятор немного ругается
 // было лень менять все на нечто нормальное
-#include "src/jacobi.h"
+#include "jacobi.h"
 
 /*
  * теперь здесь ничего не будет
